kpatch-build: fix USERSRCDIR builds
Commit 69e71f8dcc6f ("kpatch-build: cleanup kernel file backup/restore")
consolidated a bunch of kernel-tree copy and restoring.  As part of that
effort, when kpatch-build is invoked with a -s|--sourcedir USERSRCDIR
value the vmlinux file is now saved to "$TEMPDIR/kernel-backup/" and not
simply "$TEMPDIR/".  This results in kpatch-build confusion like:

  readelf: /home/jolawren/.kpatch/tmp/vmlinux: Error: No such file

Update the VMLINUX reassignment in this case to point to the new path.

Fixes: 69e71f8dcc6f ("kpatch-build: cleanup kernel file backup/restore")
Signed-off-by: Joe Lawrence <joe.lawrence@redhat.com>